python/requirements.txt
confluent-kafka==2.2.0
kafka-python==2.0.2
